requirements-cpu.txt
####################################################################
# Requirements for anaconda environment
# Run it with:
# conda install --file requirements.txt -c pytorch -c dglteam -c conda-forge
####################################################################

biopython==1.76
matplotlib
numpy
schedule
tqdm
scikit-learn
scipy

dgl==0.4.3
pytorch==1.4.0
torchvision
ignite==0.4.2
captum

Flask==1.1.2
flask-cors
requests
uwsgi
